completely rewrite authentication system
you remember that old authentication system? Its fucking gone.  Stop
looking for it, it ain't coming back.  Don't ever under any
circumstances revert this commit.  It shall live in infamy.  It was
the commit that changed a terribly bush league non-authentication
system to one that might actually work for a day or two.